Fixes the exporting of multiple classes in Webpack and further restructuring to the framework for ES6.
