+++ demos/helicopter_game/index.html
@@ -15,7 +15,8 @@
   </body>
 
   <!-- Include Pentagine -->
-  <script src="../../build/pentagine.js" type="text/javascript"></script>
+  <script src="../../build/Game.js" type="text/javascript"></script>
+  <script src="../../build/sprite.js" type="text/javascript"></script>
 
   <!-- Include game classes -->
   <script src="PlayState.js" type="text/javascript"></script>
