+++ lib/Camera.js
@@ -1,0 +1,12 @@
+class Camera {
+  constructor(x, y, width, height) {
+    this.x = x;
+    this.y = y;
+    this.width = width;
+    this.height = height;
+  }
+
+  follow() { }
+}
+
+module.exports = Camera;
