--- lib/entry.js
@@ -1,6 +1,0 @@
-var pentagine = require('./pentagine.js');
-
-console.log(pentagine);
-
-module.exports = pentagine.default;
-module.exports.Sprite = pentagine.Sprite;
